+++ shalya/__init__.py
@@ -1,1 +1,7 @@
 __version__ = "0.0.6"
+
+from . import core, host, skills, tools
+from .core import *
+from .host import *
+from .skills import *
+from .tools import *